Separate all QuestionPresenter components
