+++ src/components/popups/PopupInlineQuestionInsert.tsx
@@ -4,8 +4,8 @@ import React from "react";
 import styles from "../../styles/editor.module.css";
 import { Popup, PopupCloseContext, PopupRef } from "./Popup";
 import { Alert, Button, Container, Input, Label } from "reactstrap";
-import { InlineQuestionContext } from "../semantic/presenters/questions/questionPresenters";
 import { generateGuid } from "../../utils/strings";
+import { InlineQuestionContext } from "../semantic/presenters/questions/InlineQuestionPresenter";
 
 
 export const PopupInlineQuestionInsert = ({wide, codemirror}: { wide?: boolean, codemirror: RefObject<ReactCodeMirrorRef> }) => {
