+++ src/components/semantic/presenters/FigurePresenter.tsx
@@ -15,8 +15,8 @@ import {NON_STATIC_FIGURE_FLAG} from "../../../isaac/IsaacTypes";
 import {Alert, Input, Label} from "reactstrap";
 import { DropZoneQuestionContext } from "./questions/ItemQuestionPresenter";
 import { FigureRegionModal } from "../../FigureRegionModal";
-import { InlineQuestionContext } from "./questions/questionPresenters";
 import { isDefined } from "../../../utils/types";
+import { InlineQuestionContext } from "./questions/InlineQuestionPresenter";
 
 function githubURLFromGithubData(data: {download_url: string}, svgView?: string | null) {
     // If there is an SVG view, include at the end of the URL
