+++ src/components/semantic/presenters/ListChildrenPresenter.tsx
@@ -13,7 +13,8 @@ import {ContentType, PresenterProps} from "../registry";
 import styles from "../styles/semantic.module.css";
 import {ChildTypeOverride} from "../props/listProps";
 import {DndChoiceItemInserter, ItemChoiceItemInserter} from "./questions/ItemQuestionPresenter";
-import {CoordinateChoiceItemInserter, InlinePartInserter} from "./questions/questionPresenters";
+import { CoordinateChoiceItemInserter } from "./questions/CoordinateQuestionPresenter";
+import { InlinePartInserter } from "./questions/InlineQuestionPresenter";
 
 export interface InserterProps {
     insert: (index: number, newContent: Content) => void;
