+++ src/components/semantic/presenters/questions/questionPresenters.tsx
@@ -1,39 +1,21 @@
-import React, {createContext, useCallback, useContext, useEffect, useState} from "react";
-import {EditableDimensionalDocProp, EditableDocPropFor, EditableIDProp, EditableTitleProp} from "../../props/EditableDocProp";
+import React, {useState} from "react";
+import {EditableIDProp, EditableTitleProp} from "../../props/EditableDocProp";
 import styles from "../../styles/question.module.css";
-import {Alert, Button, Dropdown, DropdownItem, DropdownMenu, DropdownToggle} from "reactstrap";
+import {Alert, Dropdown, DropdownItem, DropdownMenu, DropdownToggle} from "reactstrap";
 import {
     AnyQuestion,
     Content,
-    IsaacCoordinateQuestion,
-    IsaacGraphSketcherQuestion,
-    IsaacInlinePart,
-    IsaacInlineQuestion,
-    IsaacMultiChoiceQuestion,
     IsaacNumericQuestion,
     IsaacQuestionBase,
-    IsaacQuickQuestion,
-    IsaacStringMatchQuestion,
-    IsaacSymbolicChemistryQuestion,
-    IsaacSymbolicQuestion,
-    PositionableFigureRegionProps,
-    Quantity
+    IsaacQuickQuestion
 } from "../../../../isaac-data-types";
 import {SemanticDocProp} from "../../props/SemanticDocProp";
-import {EditableText} from "../../props/EditableText";
-import {CheckboxDocProp} from "../../props/CheckboxDocProp";
 import {PresenterProps} from "../../registry";
 import {SemanticListProp} from "../../props/listProps";
 import {NumberDocPropFor} from "../../props/NumberDocPropFor";
 import {ChoicesPresenter} from "../ChoicesPresenter";
-import {InserterProps} from "../ListChildrenPresenter";
-import { ContentValueOrChildrenPresenter } from "../ContentValueOrChildrenPresenter";
-import { InlinePartsPresenter } from "../InlinePartsPresenter";
-import { EditableInlineTypeProp, INLINE_TYPES } from "./InlineQuestionTypePresenter";
-import { isAda } from "../../../../services/site";
-import { extractValueOrChildrenText } from "../../../../utils/content";
-import { inlineQuestionRegex } from "../../../../isaac/IsaacTypes";
-import { HUMAN_QUESTION_TYPES, QUESTION_TYPES } from "../../../../services/constants";
+import {isAda} from "../../../../services/site";
+import {HUMAN_QUESTION_TYPES, QUESTION_TYPES} from "../../../../services/constants";
 
 export const QuestionContext = React.createContext<Content | null>(null);
 
@@ -162,14 +144,6 @@ export function AnswerPresenter(props: PresenterProps<IsaacQuickQuestion>) {
     return <SemanticDocProp {...props} prop={"answer"} name="Answer" />;
 }
 
-export function QuickQuestionPresenter(props: PresenterProps) {
-    const question = props.doc as IsaacQuickQuestion;
-    return <>
-        <QuestionMetaPresenter {...props} />
-        <CheckboxDocProp doc={question} update={props.update} prop="showConfidence" label="Show confidence question" />
-    </>;
-}
-
 export function QuestionFooterPresenter(props: PresenterProps<IsaacQuestionBase>) {
     return <>
         <ChoicesPresenter {...props} />
@@ -185,379 +159,8 @@ export function HintsPresenter(props: PresenterProps<IsaacQuestionBase>) {
     return <SemanticListProp {...props} prop="hints" type="hints" />;
 }
 
-export function MultipleChoiceQuestionPresenter({showMeta = true, ...props}: {showMeta?: boolean} & PresenterProps) {
-    const {doc, update} = props;
-    const question = doc as IsaacMultiChoiceQuestion;
-    return <>
-        {showMeta && <QuestionMetaPresenter {...props} />}
-        <CheckboxDocProp doc={question} update={update} prop="randomiseChoices" label="Randomise Choices" checkedIfUndefined={true} />
-    </>;
-}
-
-const EditableSignificantFiguresMin = NumberDocPropFor<IsaacNumericQuestion>("significantFiguresMin", {label: "from", block: true});
-const EditableSignificantFiguresMax = NumberDocPropFor<IsaacNumericQuestion>("significantFiguresMax", {label: "to", block: true});
-const EditableAvailableUnits = ({doc, update}: PresenterProps<IsaacNumericQuestion>) => {
-    return <EditableText
-        onSave={(newText) => {
-            update({
-                ...doc,
-                availableUnits: newText?.split("|").map(unit => unit.trim()),
-            });
-        }}
-        text={doc.availableUnits?.join(" | ")}
-        placeHolder="Enter list of units here (|-separated)"
-        label="Available units"
-        block
-    />;
-};
-const EditableDisplayUnit = EditableDocPropFor<IsaacNumericQuestion>("displayUnit",  {label: "Display unit", block: true, format: "latex", previewWrapperChar: "$"});
-
-export function NumericQuestionPresenter({showMeta = true, ...props}: {showMeta?: boolean} & PresenterProps) {
-    const {doc, update} = props;
-    const question = doc as IsaacNumericQuestion;
-
-    return <>
-        {showMeta && <QuestionMetaPresenter {...props} />}
-        <div>
-            <CheckboxDocProp doc={question} update={update} prop="disregardSignificantFigures" label="Exact answers only" />
-        </div>
-        {!question.disregardSignificantFigures && <div className={styles.questionLabel}>
-            Significant figures:
-            <div className="row">
-                <div className="col col-lg-5">
-                    <EditableSignificantFiguresMin doc={question} update={update} />
-                </div>
-                <div className="col col-lg-5">
-                    <EditableSignificantFiguresMax doc={question} update={update} />
-                </div>
-            </div>
-        </div>}
-        <div>
-            <CheckboxDocProp doc={question} update={newQuestion => {
-                if (newQuestion.requireUnits) {
-                    delete newQuestion.displayUnit;
-                } else {
-                    delete newQuestion.availableUnits;
-                    newQuestion.choices?.forEach(choice => {
-                        if (choice.type === "quantity") {
-                            delete (choice as Quantity).units;
-                        }
-                    });
-                }
-                update(newQuestion);
-            }} prop="requireUnits" label="Require choice of units" />
-        </div>
-        {question.requireUnits ?
-            <EditableAvailableUnits doc={question} update={update} />
-            :   <EditableDisplayUnit doc={question} update={update} />}
-        <div className={styles.questionLabel} /> {/* For spacing */}
-    </>;
-}
-
-export const CoordinateQuestionContext = createContext<{numberOfCoordinates?: number, numberOfDimensions?: number}>(
-    {}
-);
-const EditableNumberOfCoordinates = NumberDocPropFor<IsaacCoordinateQuestion>("numberOfCoordinates", {label: "Number of coordinates", block: true});
-const EditableDimensions = NumberDocPropFor<IsaacCoordinateQuestion>("numberOfDimensions", {label: "Dimensions", block: true});
-const EditableSeparator = EditableDocPropFor<IsaacCoordinateQuestion>("separator", {label: "Separator", block: true});
-const EditableButtonText = EditableDocPropFor<IsaacCoordinateQuestion>("buttonText", {label: "\"Add coordinate\" button text override", block: true});
-const EditablePlaceholderValuesProp = EditableDimensionalDocProp<IsaacCoordinateQuestion>("placeholderValues");
-const EditablePrefixesProp = EditableDimensionalDocProp<IsaacCoordinateQuestion>("prefixes");
-const EditableSuffixesProp = EditableDimensionalDocProp<IsaacCoordinateQuestion>("suffixes");
-
-export function CoordinateQuestionPresenter(props: PresenterProps<IsaacCoordinateQuestion>) {
-    const {doc, update} = props;
-    const question = doc as IsaacCoordinateQuestion;
-
-    return <>
-        <QuestionMetaPresenter {...props} />
-        <EditableNumberOfCoordinates {...props} />
-        <CheckboxDocProp {...props} prop="ordered" label="Require that order of coordinates in choice and answer are the same" />
-        <EditableDimensions {...props} />
-        <CheckboxDocProp {...props} prop="useBrackets" checkedIfUndefined label="Show brackets around coordinates" />
-        <EditableSeparator {...props} />
-        <EditableButtonText {...props} />
-        <div className={styles.questionLabel}>
-            Coordinate labels:<br/>
-            <small><em>Placeholders do not accept latex. Please use a unicode equivalent such as Ψ₁.</em></small>
-            <div>
-                {[...Array(question.numberOfDimensions)].map((_, i) => 
-                    <div className={"mb-3"} key={i}>
-                        <EditablePrefixesProp {...props} dimension={i} label={"Prefix ".concat((i+1).toString())} />
-                        <span className="mx-2"/>
-                        <EditablePlaceholderValuesProp {...props} dimension={i} label={"Placeholder ".concat((i+1).toString())} />
-                        <span className="mx-2"/>
-                        <EditableSuffixesProp {...props} dimension={i} label={"Suffix ".concat((i+1).toString())} />
-                    </div>
-                )}
-            </div>
-            <div>
-                <CheckboxDocProp doc={question} update={update} prop="disregardSignificantFigures" label="Exact answers only" />
-            </div>
-            Significant figures (affects all values):
-            <div className="row mt-2">
-                <div className="col col-lg-5">
-                    <EditableSignificantFiguresMin {...props} />
-                </div>
-                <div className="col col-lg-5">
-                    <EditableSignificantFiguresMax {...props} />
-                </div>
-            </div>
-        </div>
-    </>;
-}
-
-export function CoordinateQuestionFooterPresenter(props: PresenterProps<IsaacCoordinateQuestion>) {
-    const question = props.doc as IsaacCoordinateQuestion;
-
-    return <CoordinateQuestionContext.Provider value={{
-        numberOfCoordinates: question.numberOfCoordinates,
-        numberOfDimensions: question.numberOfDimensions
-    }}>
-        <QuestionFooterPresenter {...props} />
-    </CoordinateQuestionContext.Provider>;
-}
-
-export function CoordinateChoiceItemInserter({insert, position, lengthOfCollection}: InserterProps) {
-    const numberOfCoordinates = useContext(CoordinateQuestionContext).numberOfCoordinates;
-    if (position !== lengthOfCollection) {
-        return null; // Only include an insert button at the end.
-    }
-    if (numberOfCoordinates !== undefined && lengthOfCollection >= numberOfCoordinates) {
-        return null; // Max items reached in choice
-    }
-    return <Button className={styles.itemsChoiceInserter} color="primary" onClick={() => {
-        insert(position, {type: "coordinateItem"});
-    }}>Add</Button>;
-}
-
-export function InlinePartInserter({insert, insertMultiple, position, collection, lengthOfCollection}: InserterProps) {
-    const inlineContext = useContext(InlineQuestionContext);
-    
-    const addPart = useCallback((idSuffix: string | undefined) => {
-        const id = idSuffix ? `inline-question:${idSuffix}` : undefined;
-        insert(position, {type: "inlineQuestionPart", choices: [], id: id} as IsaacInlinePart);
-    }, [insert, position]);
-
-    inlineContext.addPart = addPart;
-
-    const addMultipleParts = useCallback((ids: string[]) => {
-        // note that ids must be genuine strings, as there may only exist one undefined part
-        insertMultiple(ids.map((id, index) => [position + index, {type: "inlineQuestionPart", choices: [], id: `inline-question:${id}`} as IsaacInlinePart]));
-    }, [insertMultiple, position]);
-
-    const calculatedPartsFromContent = [
-        ...(inlineContext.textContent ? inlineContext.textContent.matchAll(inlineQuestionRegex).map(match => match.groups?.id as string) : []),
-        ...(inlineContext.figureMap ? Object.values(inlineContext.figureMap).map(pdzp => pdzp.map(dz => dz.id)).flat() : []),
-    ];
-    const unusedParts = calculatedPartsFromContent.filter(id => id && !(collection?.map(part => part.id ? part.id.slice("inline-question:".length) : undefined)?.includes(id)));
-
-    if (position !== lengthOfCollection) {
-        return null; // Only include an insert button at the end.
-    }
-
-    return <div className={styles.inlineItemsChoiceInserter}>
-        <Button color="primary" onClick={() => addMultipleParts(unusedParts)} disabled={lengthOfCollection === calculatedPartsFromContent.length}>Detect and add missing inline regions</Button>
-        <Button color="secondary" onClick={() => addPart(undefined)} disabled={collection?.some(part => part.id === undefined)}>Add new inline question part</Button>
-    </div>;
-}
-
-export function GraphSketcherQuestionPresenter(props: PresenterProps<IsaacGraphSketcherQuestion>) {
-    const {doc, update} = props;
-    const question = doc as IsaacCoordinateQuestion;
-
-    const NumCurvesInput = NumberDocPropFor<IsaacGraphSketcherQuestion>("maxNumCurves", {label: "Max number of curves", block: true});
-    const EditableAxisLabelX = EditableDocPropFor<IsaacGraphSketcherQuestion>("axisLabelX", {label: "X-axis", block: true, format: "latex", previewWrapperChar: "$"});
-    const EditableAxisLabelY = EditableDocPropFor<IsaacGraphSketcherQuestion>("axisLabelY", {label: "Y-axis", block: true, format: "latex", previewWrapperChar: "$"});
-
-    return <>
-        <QuestionMetaPresenter {...props} />
-        <div className={styles.questionLabel}>
-            Axis labels:
-            <div className="row">
-                <div className="col col-lg-5">
-                    <EditableAxisLabelX doc={question} update={update} />
-                </div>
-                <div className="col col-lg-5">
-                    <EditableAxisLabelY doc={question} update={update} />
-                </div>
-            </div>
-            <div className="row">
-                <div className="col col-lg-5">
-                    <NumCurvesInput doc={question} update={update} />
-                </div>
-            </div>
-        </div>
-    </>;
-}
-
-const EditableAvailableSymbols = ({doc, update}: PresenterProps<IsaacSymbolicQuestion>) => {
-    return <EditableText
-        onSave={(newText) => {
-            update({
-                ...doc,
-                availableSymbols: newText?.split(",").map(unit => unit.trim()),
-            });
-        }}
-        text={doc.availableSymbols?.map(unit => unit.trim()).join(", ")}
-        placeHolder="Enter list of symbols here (,-separated)"
-        label="Available symbols"
-        format={"latex"}
-    />;
-};
-const EditableFormulaSeed = EditableDocPropFor<IsaacSymbolicQuestion>("formulaSeed", {format: "latex", label: "Formula seed", placeHolder: "Enter initial state here"});
-
-const availableMetaSymbols: [string,string][] = [
-    ["_trigs", "Trigs"],
-    ["_1/trigs", "1/Trigs"],
-    ["_inv_trigs", "Inv Trigs"],
-    ["_inv_1/trigs", "Inv 1/Trigs"],
-    ["_hyp_trigs", "Hyp Trigs"],
-    ["_inv_hyp_trigs", "Inv Hyp Trigs"],
-    ["_logs", "Logarithms"],
-    ["_no_alphabet", "No Alphabet"]
-];
-
-const availableChemistryMetaSymbols: [string,string][]  = [
-    ["_state_symbols", "State Symbols"], 
-    ["_plus", "Plus"],
-    ["_minus", "Minus"],
-    ["_fraction", "Fraction"],
-    ["_right_arrow", "Right Arrow"],
-    ["_equilibrium_arrow", "Equilibrium Arrow"],
-    ["_brackets_round", "Round Brackets"],
-    ["_brackets_square", "Square Brackets"],
-    ["_dot", "Dot"]
-];  
-
-function hasSymbol(availableSymbols: string[] | undefined, symbol: string) {
-    return availableSymbols?.find(s => s === symbol);
-}
-
-function SymbolicMetaSymbols({doc, update, metaSymbols}: PresenterProps<IsaacSymbolicQuestion> & {metaSymbols: [string, string][]}) {
-    function toggle(symbol: string) {
-        const availableSymbols = [...doc.availableSymbols ?? []];
-        const index = availableSymbols.indexOf(symbol);
-        if (index !== -1) {
-            availableSymbols.splice(index, 1);
-        } else {
-            availableSymbols.push(symbol);
-        }
-        update({
-            ...doc, availableSymbols
-        });
-    }
-
-    return <div className={styles.symbolicMetaButtons}>
-        {metaSymbols.map(([symbol, label]) =>
-            <Button key={symbol}
-                size="sm"
-                color={hasSymbol(doc.availableSymbols, symbol) ? "primary" : "secondary"}
-                onClick={() => toggle(symbol)}>
-                {label}
-            </Button>
-        )}
-    </div>;
-}
-
-function SymbolicQuestionPresenterHead(props: PresenterProps<IsaacSymbolicQuestion>) {
-    return <>
-        <QuestionMetaPresenter {...props} />
-        <div className={styles.editableFullwidth}>
-            <EditableAvailableSymbols {...props} />
-        </div>
-    </>;
-}
-
-export function SymbolicChemistryQuestionPresenter(props: PresenterProps<IsaacSymbolicChemistryQuestion>) {
-    return <>
-        <SymbolicQuestionPresenterHead {...props}/>
-        {!props.doc.isNuclear && <SymbolicMetaSymbols {...props} metaSymbols={availableChemistryMetaSymbols} />}
-        <hr />
-        <CheckboxDocProp {...props} className="d-inline-block" prop="isNuclear" label="Nuclear question" />
-        <CheckboxDocProp {...props} className="d-inline-block" prop="allowPermutations" label="Allow molecule permutations" disabled={props.doc.isNuclear} />
-        <CheckboxDocProp {...props} className="d-inline-block" prop="allowScalingCoefficients" label="Allow coefficient scaling" disabled={props.doc.isNuclear} />
-        <CheckboxDocProp {...props} prop="showInequalitySeed" label="Show seed in inequality before opening the editor" />
-        <div className={styles.editableFullwidth}>
-            <EditableFormulaSeed {...props}/>
-        </div>
-    </>;
-}
-
-export function SymbolicQuestionPresenter(props: PresenterProps<IsaacSymbolicQuestion>) {
-    return <>
-        <SymbolicQuestionPresenterHead {...props} />
-        {props.doc.type === "isaacSymbolicQuestion" && <SymbolicMetaSymbols {...props} metaSymbols={availableMetaSymbols} />}
-        <hr />
-        <div className={styles.editableFullwidth}>
-            <EditableFormulaSeed {...props}/>
-        </div>
-    </>;
-}
-
-export function StringMatchQuestionPresenter(props: PresenterProps<IsaacStringMatchQuestion>) {
-    return <>
-        <QuestionMetaPresenter {...props} />
-        <CheckboxDocProp {...props} prop="multiLineEntry" label="Multi-line" />
-    </>;
-}
-
-const getInlineQuestionPresenter = (type: INLINE_TYPES, props: PresenterProps<IsaacInlinePart>) : Exclude<React.ReactNode, undefined> => {
-    switch (type) {
-        case "isaacNumericQuestion":
-            return <>
-                <hr/>
-                <NumericQuestionPresenter {...props} showMeta={false} />
-            </>;
-        case "isaacMultiChoiceQuestion":
-            return <>
-                <hr/>
-                <MultipleChoiceQuestionPresenter {...props} showMeta={false} />
-            </>;
-        case "isaacRegexMatchQuestion":
-        case "isaacStringMatchQuestion":
-            return null;
-    }
-};
-
-export function InlineQuestionPartPresenter(props: PresenterProps<IsaacInlinePart>) {
-    const [isDisabled, setIsDisabled] = useState(false);
-    const choices = <ChoicesPresenter {...props} />;
-    
-    useEffect(() => {
-        setIsDisabled(choices.props.doc.choices && choices.props.doc.choices.length > 0);
-    }, [choices.props.doc.choices]);
-
-    return <>
-        <h4><EditableTitleProp {...props} placeHolder="Question part title"/></h4>
-        <h6><EditableIDProp {...props} label="Question ID"/></h6>
-        {props.doc.id && props.doc.id.match(/^\[|\]$/) && <p className="text-danger"><i>Warning: the ID should not include the surrounding square brackets!</i></p>}
-        <EditableInlineTypeProp {...props} disabled={isDisabled} />
-        <em>Note: you cannot change the question type if any choices exist.</em>
-        {props.doc.type && getInlineQuestionPresenter(props.doc.type as INLINE_TYPES, props)}
-        {choices}
-        <SemanticDocProp {...props} prop="defaultFeedback" name="Default Feedback" />
-        <AnswerPresenter {...props} />
-    </>;
-}
-
-export const InlineQuestionContext = createContext<{
-    isInlineQuestion?: boolean,
-    addPart?: (id: string) => void,
-    figureMap?: {[figureId: string]: PositionableFigureRegionProps[]}
-    setFigureMap?: React.Dispatch<React.SetStateAction<{[figureId: string]: PositionableFigureRegionProps[]}>>,
-    textContent?: string;
-}>({});
-
-export function InlineRegionPresenter(props: PresenterProps<IsaacInlineQuestion>) {
-    const [figureMap, setFigureMap] = useState<{[figureId: string]: PositionableFigureRegionProps[]}>({});
-    const textContent = extractValueOrChildrenText(props.doc);
-    return <InlineQuestionContext.Provider value={{isInlineQuestion: true, figureMap, setFigureMap, textContent}}>
-        <h6><EditableIDProp {...props} label="Question ID"/></h6>
-        <ContentValueOrChildrenPresenter {...props} />
-        <InlinePartsPresenter {...props} />
-    </InlineQuestionContext.Provider>;
-}
+export const EditableSignificantFiguresMin = NumberDocPropFor<IsaacNumericQuestion>("significantFiguresMin", {label: "from", block: true});
+export const EditableSignificantFiguresMax = NumberDocPropFor<IsaacNumericQuestion>("significantFiguresMax", {label: "to", block: true});
 
 export function FreeTextQuestionInstructions() {
     return <div>
